\name{perry-package}
\alias{perry-package}
\docType{package}
\title{
\packageTitle{perry}
}
\description{
\packageDescription{perry}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{perry}
\packageIndices{perry}
}
\author{
\packageAuthor{perry}

Maintainer: \packageMaintainer{perry}
}
\keyword{package}
