% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.perry}
\alias{summary.perry}
\alias{summary.perrySelect}
\alias{summary.perryTuning}
\title{Summarize resampling-based prediction error results}
\usage{
\method{summary}{perry}(object, ...)

\method{summary}{perrySelect}(object, ...)

\method{summary}{perryTuning}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"perry"} or
\code{"perrySelect"} that contains prediction error results (note that the
latter includes objects of class \code{"perryTuning"}).}

\item{\dots}{currently ignored.}
}
\value{
An object of class \code{"summary.perry"}, \code{"summary.perrySelect"} or
\code{"summary.perryTuning"}, depending on the class of \code{object}.
}
\description{
Produce a summary of resampling-based prediction error results.
}
\seealso{
\code{\link{perryFit}}, \code{\link{perrySelect}},
\code{\link{perryTuning}}, \code{\link{summary}}
}
\author{
Andreas Alfons
}
\keyword{utilities}
